pyproject.toml
[project]
name = "mudd"
version = "0.1.0"
description = "Multi User Dungeon (Discord)"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "discord-py>=2.6.4",
    "python-dotenv>=1.2.1",
    "valkey-glide>=2.2.3",
]

[dependency-groups]
dev = [
    "lefthook>=2.0.13",
    "ruff>=0.14.10",
    "ty>=0.0.8",
]

[tool.ruff.lint]
select = [
    # pycodestyle
    "E",
    # Pyflakes
    "F",
    # pyupgrade
    "UP",
    # flake8-bugbear
    "B",
    # flake8-simplify
    "SIM",
    # isort
    "I",
]
